+++ src/libfetchers/tarball-cache.hh
@@ -1,0 +1,17 @@
+#pragma once
+///@file
+
+#include "ref.hh"
+#include "git-utils.hh"
+
+namespace nix::fetchers {
+
+struct TarballInfo
+{
+    Hash treeHash;
+    time_t lastModified;
+};
+
+ref<GitRepo> getTarballCache();
+
+}
